move lsm module to tag schema region
