+++ pom.xml
@@ -121,7 +121,6 @@
         <module>trigger-api</module>
         <module>rewrite-tsfile-tool</module>
         <module>external-api</module>
-        <module>lsm</module>
         <module>schema-engine-tag</module>
     </modules>
     <!-- Properties Management -->
