+++ schema-engine-tag/src/main/java/org/apache/iotdb/db/metadata/tagSchemaRegion/config/TagSchemaConfig.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.metadata.tagSchemaRegion;
+package org.apache.iotdb.db.metadata.tagSchemaRegion.config;
 
 /** tag schema region config */
 public class TagSchemaConfig {
